モデルが黒くなるバグを修正。objファイルの色が正しく描画されるように修正

+++ Game_Application/Framework/Graphics/Resources/ShaderTypes.h
@@ -90,7 +90,7 @@ namespace Shader {
 		math::vector4f Specular;
 		math::vector4f Emission;
 		float Shininess;
-		int TextureEnable;
+		float TextureEnable;
 		float pad[2];           // 16ƒoƒCƒg‹«ŠE‚É’²®
 	};
